refactor: table of contents label component
