--- src/components/TableOfContents/KuskExtensionIcon.styled.tsx
@@ -1,7 +1,0 @@
-import {ApiOutlined as RawApiOutlined} from '@ant-design/icons';
-
-import styled from 'styled-components';
-
-export const ApiOutlined = styled(RawApiOutlined)`
-  font-size: 16px;
-`;
