+++ src/components/TableOfContents/styled.tsx
@@ -1,20 +1,8 @@
-import {Tag as RawTag} from 'antd';
-
 import styled from 'styled-components';
 
 import {GlobalScrollbarStyle} from '@utils/scrollbar';
 
 import Colors from '@styles/colors';
-
-const swaggerUIMethodsColors: {[method: string]: string} = {
-  delete: '#F93E3E',
-  get: '#61AFFE',
-  head: '#9012FE',
-  options: '#0D5AA7',
-  patch: '#50E3C2',
-  post: '#49CC90',
-  put: '#FCA130',
-};
 
 export const ContentContainer = styled.div`
   margin-top: 20px;
@@ -51,45 +39,8 @@ export const ContentLabel = styled.div<{$level: 'top' | 'path' | 'operation'; $r
   }
 `;
 
-export const LabelMethodTag = styled(RawTag)<{$deprecated: boolean; $method: string}>`
-  ${({$deprecated, $method}) => `
-    background-color: ${$deprecated ? '#EBEBEB' : swaggerUIMethodsColors[$method]};
-    opacity: ${$deprecated ? '0.6' : '1'};
-  `}
-
-  font-weight: bold;
-  color: ${Colors.grey2};
-  border: none;
-  padding: 1px 8px;
-  margin: 0;
-`;
-
-export const LabelPath = styled.span<{$deprecated: boolean}>`
-  ${({$deprecated}) => `
-    text-decoration: ${$deprecated ? 'line-through' : ''};
-    opacity: ${$deprecated ? '0.65' : '1'};
-  `}
-`;
-
-export const LabelTag = styled(RawTag)`
-  background-color: ${Colors.grey4};
-  color: ${Colors.whitePure};
-  margin: 0;
-  font-size: 11px;
-`;
-
 export const TableOfContentsContainer = styled.div`
   margin-bottom: 30px;
-`;
-
-export const TableOfContentsLabel = styled.div<{$level: 'top' | 'path' | 'operation'}>`
-  ${({$level}) => `
-    margin-bottom: ${$level === 'operation' ? '14px' : '0px'};
-  `}
-
-  display: flex;
-  align-items: center;
-  gap: 12px;
 `;
 
 export const TableOfContentsTitle = styled.span`
